requirements.txt
astroid==1.5.3
blessed==1.14.2
botocore==1.8.7
cement==2.8.2
colorama==0.3.7
cycler==0.10.0
dj-database-url==0.4.2
dj-static==0.0.6
Django==1.11.7
django-appconf==1.0.2
django-crispy-forms==1.7.0
django-imagekit==4.0.2
django-js-asset==0.1.1
django-markdown-deux==1.0.5
django-pagedown==0.1.3
docker-py==1.7.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
gunicorn==19.7.1
isort==4.2.15
jmespath==0.9.3
lazy-object-proxy==1.3.1
markdown2==2.3.5
mccabe==0.6.1
olefile==0.44
pathspec==0.5.0
pilkit==2.0
Pillow==4.3.0
psycopg2==2.7.3.2
pylint==1.7.4
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.3
PyYAML==3.12
requests==2.9.1
semantic-version==2.5.0
six==1.11.0
static3==0.7.0
tabulate==0.7.5
termcolor==1.1.0
wcwidth==0.1.7
websocket-client==0.44.0
wrapt==1.10.11
